fix: get available rooms
