+++ apps/busca-hoteis/package.json
@@ -24,6 +24,7 @@
     "@types/node": "^20.12.7",
     "tsup": "^8.0.2",
     "tsx": "^4.7.3",
-    "typescript": "^5.4.5"
+    "typescript": "^5.4.5",
+    "@repo/tsconfig": "workspace:*"
   }
 }
