+++ apps/busca-hoteis/tsconfig.json
@@ -1,19 +1,3 @@
 {
-  "compilerOptions": {
-    "target": "es2017",
-    "module": "ES2022",
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "strictPropertyInitialization": false,
-    "moduleResolution": "Bundler",
-    "resolveJsonModule": true,
-    "strict": true,
-    "skipLibCheck": true,
-    "baseUrl": ".",
-    "paths": {
-      "~/*": ["./src/*"]
-    }
-  },
-  "exclude": ["node_modules"],
-  "include": ["src"]
+  "extends": "@repo/tsconfig/base.json"
 }
