+++ packages/tsconfig/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@repo/tsconfig",
+  "private": true
+}
